+++ yampt.translator/source/dialog/settings/translator_settings_dialog.cpp
@@ -35,12 +35,10 @@ translator_settings_dialog_t::translator_settings_dialog_t(
 	m_category_list->addItem("Appearance");
 	m_category_list->addItem("Shortcuts");
 	m_category_list->addItem("Language");
-	m_category_list->addItem("Translation");
 
 	m_content_stack->addWidget(m_appearance_view);
 	m_content_stack->addWidget(m_shortcuts_view);
 	m_content_stack->addWidget(m_language_view);
-	m_content_stack->addWidget(m_translation_view);
 
 	connect(m_category_list, &QListWidget::currentRowChanged, m_content_stack, &QStackedWidget::setCurrentIndex);
 
@@ -87,7 +85,6 @@ translator_settings_dialog_t::translator_settings_dialog_t(
 
 	m_appearance_view->load(m_settings);
 	m_language_view->load(m_settings);
-	m_translation_view->load(m_settings);
 	m_shortcuts_view->load(m_settings);
 	m_category_list->setCurrentRow(0);
 }
@@ -96,7 +93,6 @@ void translator_settings_dialog_t::apply_all()
 {
 	m_appearance_view->save(m_settings);
 	m_language_view->apply(m_settings);
-	m_translation_view->apply(m_settings);
 	m_shortcuts_view->apply(m_settings);
 	m_settings.sync();
 }